improved code quality, fixed incomplete bug test, added chained methods, added new BDD methods
