+++ .gitignore
@@ -1,3 +1,4 @@
 vendor
 .idea
+.phpunit.result.cache
 composer.phar
